Move hover style from _base.scss to _layout.scss
iOS will activate the hover state, if present, the first time
an element is tapped, meaning two taps are required to perform
an action like following a link.

This commit moves the hover styles to the desktop-only
stylesheet so iOS never even knows they’re there.

Some concerns:

 - Still requires two taps on iPads (and iPhones in landscape)
 - This lift-and-shift of the styles doesn’t seem like a great solution
   but my Sassiness isn’t sassy enough to know if there’s a better way.
 - There are probably cobrand base.scss files with :hover styles too

Connects to #2294.